added images for loading the help dict, cleared up the mess with loading dict task cancelling. cleared the mess with switching languages. Fixed a newly introduced typo bug in learn fragment

+++ src/com/learnit/LearnIt/controllers/AddWordsController.java
@@ -120,7 +120,7 @@ public class AddWordsController implements
 	}
 
 	@Override
-	public void onProgressUpdate(Integer... values) {
+	public void onProgressUpdate(Double... values) {
 
 	}
